+++ tests/testthat/test_wa_data.R
@@ -26,6 +26,29 @@ test_that("wa_features", {
   )
 })
 
+test_that("wa_carbon", {
+  wa_carbon <- get_wa_carbon()
+  expect_is(wa_carbon, "SpatRaster")
+  expect_equal(terra::nlyr(wa_carbon), 1)
+  expect_true(
+    all(terra::global(wa_carbon, "min", na.rm = TRUE)[[1]] >= 0)
+  )
+})
+
+test_that("wa_attr", {
+  wa_attr <- get_wa_attr()
+  expect_is(wa_attr, "tbl_df")
+  expect_named(
+    wa_attr,
+    c(
+      "feature", "binomial", "family", "order",
+      "extinction_prob", "interest_score"
+    )
+  )
+  expect_equal(wa_attr$feature, names(get_wa_features()))
+  expect_equal(nrow(wa_attr), nrow(na.omit(wa_attr)))
+})
+
 test_that("locked data are feasible", {
   wa_locked_in <- get_wa_locked_in()
   wa_locked_out <- get_wa_locked_out()
@@ -53,9 +76,11 @@ test_that("standardized missing values", {
   wa_locked_in <- get_wa_locked_in()
   wa_locked_out <- get_wa_locked_out()
   wa_features <- get_wa_features()
+  wa_carbon <- get_wa_carbon()
   expect_equal(nonNA_cells(wa_pu), nonNA_cells(wa_locked_in))
   expect_equal(nonNA_cells(wa_pu), nonNA_cells(wa_locked_out))
   expect_equal(nonNA_cells(wa_pu), nonNA_cells(sum(wa_features)))
+  expect_equal(nonNA_cells(wa_pu), nonNA_cells(wa_carbon))
 })
 
 test_that("rasters are comparable", {
@@ -63,6 +88,7 @@ test_that("rasters are comparable", {
   wa_locked_in <- get_wa_locked_in()
   wa_locked_out <- get_wa_locked_out()
   wa_features <- get_wa_features()
+  wa_carbon <- get_wa_carbon()
   expect_true(
     terra::compareGeom(wa_pu, wa_locked_out, stopOnError = FALSE)
   )
@@ -71,5 +97,8 @@ test_that("rasters are comparable", {
   )
   expect_true(
     terra::compareGeom(wa_pu, wa_features, stopOnError = FALSE)
+  )
+  expect_true(
+    terra::compareGeom(wa_pu, wa_carbon, stopOnError = FALSE)
   )
 })